+++ plone/app/contenttypes/behaviors/leadimage.py
@@ -7,6 +7,7 @@ from plone.supermodel import model
 from zope import schema
 from zope.component import adapter
 from zope.interface import implementer
+from zope.interface import Interface
 from zope.interface import provider